Demoon

+++ Tiralabra_maven/src/main/java/com/mycompany/tiralabra_maven/domain/Sarmio.java
@@ -11,11 +11,11 @@ package com.mycompany.tiralabra_maven.domain;
  * @author szetk
  */
 public class Sarmio {
-    private int x;
-    private int y;
-    private int z;
+    private long x;
+    private long y;
+    private long z;
     
-    public Sarmio(int x, int y, int z){
+    public Sarmio(long x, long y, long z){
         this.x = x;
         this.y = y;
         this.z = z;
@@ -26,31 +26,31 @@ public class Sarmio {
         return "x = " + this.x + ", y = " + this.y + ", z = " + this.z;
     }
     
-    public int getTilavuus(){
+    public long getTilavuus(){
         return this.x * this.y * this.z;
     }
 
-    public int getX() {
+    public long getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(long x) {
         this.x = x;
     }
 
-    public int getY() {
+    public long getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(long y) {
         this.y = y;
     }
 
-    public int getZ() {
+    public long getZ() {
         return z;
     }
 
-    public void setZ(int z) {
+    public void setZ(long z) {
         this.z = z;
     }
